Backend/requirements.txt
asgiref==3.5.2
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.0
pika==1.3.0
psycopg2-binary==2.9.3
PyJWT==2.5.0
pytz==2022.2.1
sqlparse==0.4.3
django-rest-swagger==2.2.0
